move plot functions into utils
