+++ examples/text_summarization/text_summarization/evaluator.py
@@ -12,9 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from typing import List
-from typing import Optional
 from typing import Tuple
-from typing import Union
 
 import evaluate
 import numpy as np
@@ -26,10 +24,8 @@ from text_summarization.workflow import TestSample
 from text_summarization.workflow import TestSampleMetric
 from text_summarization.workflow import TestSuiteMetric
 
-from kolena.workflow import AxisConfig
-from kolena.workflow import Curve
-from kolena.workflow import CurvePlot
-from kolena.workflow import Histogram
+from examples.text_summarization.text_summarization.utils import compute_metric_vs_metric_plot
+from examples.text_summarization.text_summarization.utils import compute_score_distribution_plot
 from kolena.workflow import Plot
 from kolena.workflow.evaluator_function import EvaluationResults
 from kolena.workflow.evaluator_function import TestCases
@@ -93,64 +89,6 @@ def compute_aggregate_metrics(
         ROUGE_L=mean_metric_ignoring_failures("ROUGE_L"),
         BLEU=mean_metric_ignoring_failures("BLEU"),
         inf_to_gt_word_count=mean_metric_ignoring_failures("inf_to_gt_word_count"),
-    )
-
-
-def compute_score_distribution_plot(
-    score: str,
-    metrics: List[Union[TestSampleMetric, Inference]],
-    binning_info: Optional[Tuple[float, float, float]] = None,  # start, end, num
-    logarithmic: bool = False,
-) -> Histogram:
-    scores = [getattr(m, score) for m in metrics]
-    if logarithmic:
-        bins = np.logspace(*binning_info, base=2)
-    else:
-        bins = np.linspace(*binning_info)
-
-    hist, _ = np.histogram(scores, bins=bins)
-    return Histogram(
-        title=f"Distribution of {score}",
-        x_label=f"{score}",
-        y_label="Count",
-        buckets=list(bins),
-        frequency=list(hist),
-        x_config=AxisConfig(type="log") if logarithmic else None,
-    )
-
-
-def compute_metric_vs_metric_plot(
-    x_metric: str,
-    y_metric: str,
-    x_metrics: List[Union[TestSampleMetric, Inference]],
-    y_metrics: List[Union[TestSampleMetric, Inference]],
-    binning_info: Optional[Tuple[float, float, float]] = None,  # start, end, num
-    x_logarithmic: bool = False,
-    y_logarithmic: bool = False,
-) -> CurvePlot:
-    y_values = [getattr(m, y_metric) for m in y_metrics]
-    x_values = [getattr(m, x_metric) for m in x_metrics]
-    if x_logarithmic:
-        bins = list(np.logspace(*binning_info, base=2))
-    else:
-        bins = list(np.linspace(*binning_info))
-
-    bins_centers: List[float] = []
-    bins_values: List[float] = []
-    for i in range(len(bins) - 1):
-        lo, hi = bins[i : i + 2]
-        bin_values = [y for y, x in zip(y_values, x_values) if lo <= x < hi]
-        if len(bin_values) > 0:
-            bins_centers.append(lo + ((hi - lo) / 2))
-            bins_values.append(np.mean(bin_values))
-
-    return CurvePlot(
-        title=f"{y_metric} vs. {x_metric}",
-        x_label=f"{x_metric}",
-        y_label=f"{y_metric}",
-        curves=[Curve(x=bins_centers, y=bins_values)],
-        x_config=AxisConfig(type="log") if x_logarithmic else None,
-        y_config=AxisConfig(type="log") if y_logarithmic else None,
     )
